backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pydantic==2.9.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
